Move blogpost utils into their own folder
